fix: resolve lint violations from rebase and sync test assertion
- task_query.go: remove redundant embedded field RetrieveRequest from
  selectors (QF1008, 6 occurrences in upstream ORDER BY code)
- pipeline.go: use fmt.Fprintf instead of WriteString(fmt.Sprintf) (QF1012)
- helpers_test.go: use tagged switch instead of if-else chain (QF1003)
- task_update_test.go: remove unnecessary int64() conversion (unconvert)
- util_test.go: sync assertion to lowercase error message for ST1005

+++ internal/datanode/external/task_update_test.go
@@ -197,7 +197,7 @@ func (s *UpdateExternalTaskSuite) TestBalanceFragmentsToSegments_SingleFragment(
 	s.NoError(err)
 	s.Len(result, 1)
 	s.Equal(int64(500), result[0].GetNumOfRows())
-	s.Equal(int64(storage.StorageV3), result[0].GetStorageVersion(),
+	s.Equal(storage.StorageV3, result[0].GetStorageVersion(),
 		"external segments should have StorageVersion=V3")
 }
 
@@ -236,7 +236,7 @@ func (s *UpdateExternalTaskSuite) TestBalanceFragmentsToSegments_MultipleFragmen
 
 	// Verify all segments have StorageVersion=V3
 	for i, seg := range result {
-		s.Equal(int64(storage.StorageV3), seg.GetStorageVersion(),
+		s.Equal(storage.StorageV3, seg.GetStorageVersion(),
 			"segment %d should have StorageVersion=V3", i)
 	}
